pyproject.toml
[project]
name = "actnpc-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.2.1",
    "openai>=1.0.0",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.30.0",
]
